training/requirements.txt
# HTR training — fine-tunes TrOCR, logs to MLflow, quality gate.
#
# CPU-only torch to keep image size small (~1.5 GB total).

# Torch (CPU) — installed separately in Dockerfile from pytorch index
# torch==2.4.1   ← via pytorch CPU index, NOT pypi

transformers==4.45.2
sentencepiece==0.2.0
PyYAML==6.0.2
PyArrow==18.1.0
Pillow==10.4.0
numpy==1.26.4
minio==7.2.9
mlflow==2.17.2
